refactor: rename Notification to Event
* Adds a compatibility shim in bugsnag.notification.Notification
* Prints a warning if Notification is used
